+++ src/interrupt.rs
@@ -23,7 +23,7 @@ pub fn disable() {
 ///
 /// # Safety
 ///
-/// - In any function `f()` that calls `enable`, `CriticalSection` or `&CriticalSection` tokens cannot be used in `f()`'s body after the 
+/// - In any function `f()` that calls `enable`, `CriticalSection` or `&CriticalSection` tokens cannot be used in `f()`'s body after the
 ///   call to `enable`. If `f()` owns `CriticalSection` tokens, it is recommended to [`drop`](https://doc.rust-lang.org/nightly/core/mem/fn.drop.html)
 ///   these tokens before calling `enable`.
 #[inline(always)]
@@ -47,14 +47,15 @@ pub unsafe fn enable() {
 /// This as also known as a "critical section".
 pub fn free<F, R>(f: F) -> R
 where
-    F: FnOnce(&CriticalSection) -> R,
+    F: for<'a> FnOnce(&'a CriticalSection<'a>) -> R,
 {
     let status = ::register::sr::read();
 
     // disable interrupts
     disable();
 
-    let r = f(unsafe { &CriticalSection::new() });
+    let cs = unsafe { CriticalSection::new() };
+    let r = f(&cs);
 
     // If the interrupts were active before our `disable` call, then re-enable
     // them. Otherwise, keep them disabled