+++ src/peripheral/mod.rs
@@ -1,4 +1,0 @@
-//! Core peripherals
-//!
-
-pub use bare_metal::Peripheral;